+++ pysdd/weight_optimization_c.pxd
@@ -12,7 +12,7 @@ cimport sddapi_c
 cdef extern from "weight_optimizer.h":
 
     void optimize_weights(sddapi_c.SddNode* sdd, sddapi_c.SddManager* mgr, int m_instances,
-                          int n_optimize, int* ind_optimize, long double* weights_optimize, int* counts_optimize,
-                          int n_fix, int* ind_fix, long double* weights_fix, int* counts_fix,
+                          int n_optimize, int* ind_optimize, double* weights_optimize, int* counts_optimize,
+                          int n_fix, int* ind_fix, double* weights_fix, int* counts_fix,
                           long double prior_sigma, long double l1_const, int max_iter, long double delta,
                           long double epsilon)
